requirements.txt
# Core ML libraries (compatible with Python 3.7+)
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<1.25.0
scikit-learn>=1.0.0,<1.4.0
scipy>=1.7.0,<1.12.0

# Data processing
nltk>=3.7,<3.9
matplotlib>=3.5.0,<3.8.0
seaborn>=0.11.0,<0.13.0

# MLOps tools
mlflow>=2.0.0,<2.8.0
prefect>=2.10.0,<2.15.0
evidently>=0.4.0,<0.5.0

# Model serving
fastapi>=0.95.0,<0.105.0
uvicorn>=0.20.0,<0.25.0
pydantic>=1.10.0,<2.5.0

# Cloud simulation
localstack>=2.0.0,<2.5.0
boto3>=1.26.0,<1.29.0
moto>=4.0.0,<4.3.0

# Development and testing
pytest>=7.0.0,<7.5.0
pytest-cov>=4.0.0,<4.2.0
black>=22.0.0,<24.0.0
flake8>=5.0.0,<7.0.0
isort>=5.10.0,<5.13.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.20.0,<6.26.0

# Environment management
python-dotenv>=0.19.0,<1.1.0
pyyaml>=6.0,<6.1.0

# Additional utilities
tqdm>=4.64.0,<4.67.0
click>=8.1.0,<8.2.0
requests>=2.28.0,<2.32.0
joblib>=1.2.0,<1.4.0 